Removes Person and Phone and Updates User models to Immutable
This is a bit of an overhaul. Removes "test" models Person and Phone.
They are no longer needed as this repo will serve as a template. For
that use, a User with permission is needed instead.

Also all User models are now Immutable and uses copy instead to user the
 best practice and use of the data class.

Also renames, and add new User models. We now have 3 different models

1. User (db)
2. UserOUT (what is sent to client)
3. UserIN (what client sends to server)

This way, client now what to get and what to send. Some fields are
"read only", and set server side in UserService.kt